+++ src/postgres_mcp/explain/__init__.py
@@ -1,13 +1,7 @@
 """PostgreSQL explain plan tools."""
 
-from ..dta.artifacts import ExplainPlanArtifact
-from .tools import ErrorResult
 from .tools import ExplainPlanTool
-from .tools import JsonResult
 
 __all__ = [
-    "ErrorResult",
-    "ExplainPlanArtifact",
     "ExplainPlanTool",
-    "JsonResult",
 ]
